+++ Serendipity/src/main/java/org/shawn/games/Serendipity/Search/History/PieceToHistory.java
@@ -1,4 +1,4 @@
-package org.shawn.games.Serendipity.History;
+package org.shawn.games.Serendipity.Search.History;
 
 import com.github.bhlangonijr.chesslib.*;
 import com.github.bhlangonijr.chesslib.move.Move;
