+++ Serendipity/src/main/java/org/shawn/games/Serendipity/Search/Listener/ISearchListener.java
@@ -1,4 +1,4 @@
-package org.shawn.games.Serendipity.Search.Listeners;
+package org.shawn.games.Serendipity.Search.Listener;
 
 public interface ISearchListener
 {
